move helper functions to tox_pin_deps.compile
having executable code in `__init__.py` for the
package is irritating, since all module imports
have to go through that file.
